03_open_api/0801_team_final/requirements.txt
streamlit
pandas
numpy
matplotlib 
plotly
openpyxl
html5lib
PyMySQL
sqlalchemy
python-dotenv
streamlit-option-menu
pydeck
geopy
requests